tests/lwip: fixes the output of netif->name in ifconfig
`netif.name` is defined as `char[2]`. Therefore, the name of a netif doesn't have a terminating `\0` character. Using `%s` in `printf` works only for the first `netif` where the following member `netif.num` is zero. But it leads to garbage in the output for all other `netif` where `netif.num` is greater than 0.