coapfileserver: Cling to filesystem example
The filesystem example is ideal to demonstrate the fileserver as it has
a usable default set of demo files, and different file systems can be
mounted for experimentation.

The impact on the filesystem example outside the coapfileserver context
is minimal: Only a few lines of includes (2), static configuration (17
including comments and whitespace) and startup code (2) are added under
ifdef guards that clearly mark them as optional and Gcoap related. They
are only enabled if NETWORK=1 is set.