sifive_e: Initial include of Qemu sifive_e board
Qemu v2.12.0 supports the RISCV sifive_e machine. This machine closely
matches the hifive1 board, emulating most of the hardware of the fe310
core. One notable difference is that the clock for the mtime is running
at 10 MHz.

This commit adds the sifive_e target as a separate board. It can be used
by compiling for this board and starting qemu with `make emulate`.