cpu/stm32: Fix CPU_HAS_BITBAND
On STM32F1 `SRAM_BB_BASE` is defined, but that family does not support
bit-banding. The test of bitband support checks for this macro and assumes
correct vendor header files, which sadly is not the case for the STM32F1.

This commit works around the bug in vendor header files.