fe310: Fix potential deadlock in thread_yield_higher
The software interrupt may be raised before the WFI instruction is
reached, if it is and no other interrupt is raised afterwards the
thread will deadlock as it will never awake from the WFI.

I noticed this bug with a custom RISC-V simulator and cannot reproduce
it on real hardware though. However, even on real hardware the WFI
instruction may be implemented as a NOP, thus the current code is
suboptimal either way. The change proposed here simply busy waits until
the interrupt handler for the software interrupt is invoked (which sets
CLINT_MSIP), since this should only take a few cycles it shouldn't have
a big impact on power consumption.