boards/native: use cross-compiler for native if needed
On newer Ubuntu version (20.10) it looks like it's no longer possible
to produce 32 bit binaries with the native gcc.

The problem is that `libgcc.a` can't be found. `/usr/lib/gcc/i686-linux-gnu/10/libgcc.a`
still exists, but it's not within the search patch of `gcc -m32 -print-search-dirs`.

An easy solution is to install the `gcc-i686-linux-gnu` cross toolchain.
That means now we also need to set a cross-compile prefix also on `native`.