cpu/native: remove unused can_linux module
b4f29035ce6ee12949df1227ee463340ea8e8571 adapted the can_linux module to
the periph_can interface. This is a cleanup of some things that stayed
behind. Here the makefile is removed, the references to can_linux in the
dependency resolution and configuration Makefile are changed to the
standard periph_can, and the startup code is adapted.