gnrc/netif_bus: generate distinct events for global and link-local addresses
The `GNRC_IPV6_EVENT_ADDR_VALID event would be generated for *both* link-local
and global addresses.

So if a user wants to be notified about a global address, they have to do

    msg_bus_subscribe(&sub, GNRC_IPV6_EVENT_ADDR_VALID);
    msg_t m;
    do {
        if (xtimer_msg_receive_timeout(&m, ADDR_TIMEOUT_US) < 0) {
            return -1;
        }
    } while (ipv6_addr_is_link_local(m.content.ptr));

This is cumbersome and error-prone.

Instead, introduce a two separate events for link-local and global addresses.
Now we don't have to loop anymore until the 'proper' message is received, we
can simply do

    msg_bus_subscribe(&sub, GNRC_IPV6_EVENT_GLOBAL_ADDR_VALID);
    return xtimer_msg_receive_timeout(&m, ADDR_TIMEOUT_US);