pkg/lvgl: set LVGL thread priority higher than main
Using a lower priority than of the main thread would make
the LVGL thread to preempt the main thread, for example,
if you do a `xtimer_msleep(200);` on the main thread
giving any chance for the LVGL thread to preempt it
while working with LVGL on the main thread would
make the CPU crash.

To reproduce the bug just create various widgets, and fill
them with data from a sensor that "waits" for data to be ready.

Signed-off-by: Jean Pierre Dudey <me@jeandudey.tech>