dist/tools/renode: adapt initialization
With https://github.com/renode/renode/issues/144 merged, it is possible
to construct paths from variables. This allows us to include board and
CPU definitions for Renode with RIOT-OS.

For now, $BOARD, $CPU, $RIOTBASE, $RIOTBOARD, $RIOTCPU and $ELFFILE
should be sufficient. It can be easily extended.

This commit moves the default `board.resc` to a renode subfolder as
well, to allow for improved setups with custom board and/or CPU
definitions.