sys/oneway-malloc: export oneway_malloc()
The one-way malloc() implementation is declared weak, so it will be
overwritten if libc provides a proper implementation.

For some situations however it is useful to allocate a block of memory
on init, that is not going to be freed.
Here we don't need to pay the additional accounting overhead of `malloc()`,
a one-way malloc would be entirely sufficient.

For this, make the one-way malloc implementation accessable as `oneway_malloc()`.