mtd_spi_nor: Add support for autoconfiguration using SFDP
JEDEC has a standardized device information tables for serial flash
chips. This commit adds limited support for reading these tables and
configure the flash parameters based on these tables. A flag is added to
disable this to allow for manual configuration. A downside of this
change is that most SPI NOR flash parameters have to reside in the
MCU memory.