tests/sys_atomic_utils_unittests: spice up test
Initializing the state variables per loop with a value where some bits are set
and some are cleared makes much more sense. E.g. previously the bitwise and
applied on the state variable initialized with zero was unlikely to detect
implementation issues, as the state value never changed. So if the bitwise
and would be incorrectly implemented as a nop, it would have passed the unit
test.