nanocoap: add coap_opt_add_uri_path_with_len()
It can often be handy to append a non-0 terminated string to a path,
e.g. because only a substring should be added or the result of a
function that doesn't NULL-terminate strings (base64url_encode(), strncpy()).

Add a helper function for that case, that acts just as coap_opt_add_uri_path()
but allows to specify a length.