syslog: add basic syslog implementation
Add a partially rfc5424 compatible syslog implementation. This module
can print log messages formatted as described in rfc5424 in stdio or in
files.
The stdio backend simply prints out every log message on stdio.
The file backend prints log messages in file and can perform a basic
rotation. After a configurable number of bytes is printed, the log file
is rotated. The maximum number of files that can be kept is also
configured.
For convenience LOG_* macros can be used to print into the syslog.