makefiles: filter external_module_dirs with USEMODULE
All paths in EXTERNAL_MODULE_DIRS are added to DIRS, making them
built even if they are not activated with USEMODULE.
Worst, their dependencies are parsed and activated too, leading to
build non related modules to the original application.
To reproduce, comment "USEMODULE += external_module" in
$(RIOTBASE)/tests/external_module_dirs/Makefile: the application
build and it should not.
To fix this, filter EXTERNAL_MODULE_DIRS with USEMODULE variable.
To handle cases where module name is different from its
folder name, use sed to extract module name from Makefile and
try to find the module in USEMODULE.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>