boards/common/esp8266: Set RTS and DTR to 0 on pyterm.
When flashing the ESP8266 with the flash tool it is possible to connect
RTS to the reset line and DTR to GPIO0 to automatically reboot the
device into flash mode and then reboot the device again into normal
mode.

pyterm leaves these two signals unset unless we specify a --set-rts=0 or
1, and the value of these lines would depend on the UART driver.

This patch explicitly sets RTS and DTR to 0 (high level) in the
Makefile.include config so `make term` releases the reset line,
otherwise the device may be left in reset mode after pyterm starts.