makefiles/tools/uf2conv.inc.mk: partial riotboot support
Individual files need to be converted to uf2 format, targets
flashing individual slots or the bootloader will work:

- riotboot/flash-slot%
- riotboot/flash-bootloader

'flash' also works by flashing both the bootloader and slot0
independently.

But not targets flashing combined/extended versions since conversion
of the blob is not possible with the uf2conv.py script.