build-system: auto-locate modules by name
This commit introduces the MODULE_DIRS variable that works just like
EXTERNAL_MODULE_DIRS, but it is intended for internal modules. The
major difference is that external modules can have the same name as
an internal module, in which case the external module is used instead
of the internal one. This allows to provide drop-in replacements of
internal modules, which can be useful during development e.g. to
quickly benchmark different implementations.