tests/turo: reset before opening terminal
Some BOARD terminals are slow to start so if the BOARD is reset after
opening the terminal this can lead to off-by-one error when parsing
the output.

So instead do as testrunner and reset before opening the terminal.
Note that no interactive_sync is needed since a similar mechanism
happens in ShellInteraction.