sys/net/dhcpv6: send final RA after configuring downstream subnets
If the upstream router manages a /56 and hands our DHCPv6 IA_PD enabled
router a /62 for use in it's downstream network, hosts in the upstream
network may still assume that hosts on the new subnet are on-link.

To fix this, we include the  route information option about the subnets
routed via our router in the last router advertisement.

This means we must delay sending the last router advertisement until the
downstream interface(s) have been configured.

To speed up configuration, start with RAs disabled and only enable sending
them once we have received a valid configuration via DHCPv6.