pkg/nimble/autoconn: stop scan/adv on NETIF_ABORT_SLAVE
When a connection is aborted and the node in question is a slave
then it can be advertising/scanning at the time. The incoming
event triggers the state to be changed to IDLE but ongoing adv/scan
which will cause 'nimble_netif_conn_start_adv' to fail to allocate
a handle as the netif will still be busy. Therefore when
stop scan/stop before switching the state.