tinydtls: sock_dtls: only use ifindex with link-local addresses
For global addresses the out-going interface is typically decided by the
forwarding information base, so _requiring_ the user to set it with
`sock_dtls` is undesirable. This has the caveat, that now on
`sock_udp_send()` in the `_write()` method, the sock does not know via
which network interface to send a packet with a global address, but
again: this is really more a task for the underlying network stack, not
the sock layer.