gnrc_netif: always join solicited-node multicast group if 6LN
A border router will send a neighbor solicitation to the
solicited-node multicasts if it encounters a packet from a global address
that is not registered with the border router.

This happens when the border router is restarted while some nodes were
online.

As the nodes were not in the solicited-node multicast group, they would
drop the neighbor solicitation which resulted in them not being able to
route any packets through the border router until they would reboot themself
(or their prefix expired).

To fix this, always join the solicited-node multicast address as a 6LN.
This allows to have a connection across BR restarts.