drivers/dose: add header to DOSE frame
Setting a timeout for each byte is costly and with high symbol rates
causes missed bytes and broken frames.

To mitigate this, prepend a header before the DOSE ethernet frame that
comtains the frame length.
This also means we can drop the escaping of bytes since the length is
known in advance.

With this we only have to set two timers in the RX path:

 - one when the start-of-frame marker is received
 - one when we know the actual length of the frame and can estimate the RX time

If setting a timer takes longer than receiving a byte, we still lose with just
a single timer, but a RX FIFO can save us here a bit.
What *really* saves us is that we still set a timer for every TX byte too.

I will prepare a follow-up patch that allows to switch from expensive xtimer
to cheap countdown timers so this will also work when we speed up the TX path.

With this it is possible to use DOSE on a 1 MBaud UART between `samr21-xpro`
and `same54-xpro`.