drivers/dose: fix standby mode
Previously we would not disable the sense pin, this meant that
a node could still wake up if the standby pin is not connected.

Properly disable sensing when in standby and wait for a state
transition to IDLE to avoid aborting a reception and messing up the
DOSE internal state.