shell/gnrc_icmpv6_echo: Fix hang with no msg queue
`msg_avail()` will return -1 if the thread has no message queue.
Casting this to unsigned will result in the `ping` command to try
receiving 4294967295 messages, which hangs the shell.

Drop the cast to `unsigned` and the loop behaves as intended.
But then it's still wrong: If new messages become available, they
would be ignored.

So change the `for` loop to a `while` loop. The index variable is
not used at all.