sys/ztimer: add 'ztimer_periph_rtt_auto_select'
This is a temporary fix for Issue RIOT-OS#17060. It adds a module that
enables the auto-inclusion of 'ztimer_periph_rtt'. This module,
'ztimer_periph_rtt_auto_select' is set as a DEFAULT_MODULE and can therefore
be deselect at application level of in CPU/BOARDs 'Makefile.default'.

Limitations:
- This does not allow for modules to disable the auto-selection, e.g.:
  there can't be a depency where 'rtt_rtc' disables
  'ztimer_periph_rtt_auto_select'
- This does not disallow direct inclusions of `ztimer_periph_rt*`,
  since this only disables auto inclusion of these modules

This is a temporary solution since this is already possible with
Kconfig, but not in make.