tests/unittests: fix test vectors
In the tests fixed in this commit, the header tells that there is a PAN
ID in the header, but no PAN ID was provided in the test vector. This
amends the test vector with the PAN ID (as done in other tests in this
test suite) and adds a check if the PAN was correct.