pkg/driver_atwinc15x0: fix dependendies
The package is a helper package required for the `atwinc15x0` driver module and cannot be used standalone. Therefore it should depend on the `atwinc15x0` driver module and only be visible and force selected when the `atwinc15x0` driver module is used. The dependencies of `periph_*` are ensured by the `atwinc15x0` driver module.