core/assert: check assert at compile time, if possible
If value of the expression `expr` in `assert(expr)` is know at
compile time, this abused the optimizer and linker to fail compilation
if the expression is not true. It falls back to runtime checking if
the value is not known at compile time.

**Beware**: This will break the common pattern to add an `assert(0);`
            to label code that *should* be unreachable, as clearly the
            value of the literal `0` is known at compile time and never
            true.

For this reason, `assert_unreachable()` is added that can be used
instead. A semantic patch has been added, so that the CI will ask
contributors to replace `assert(0)` with `assert_unreachable()`.